refactor(script.js) broke script.js into multiple components
